Switched output to unicode half blocks
The screen is now drawn to the terminal with unicode half blocks so that it
is more likely to fit on most displays.  The program should now be launched
with the 'termboy' wrapper script in order to correct the aspect ratio.

All log statements have been commented out for now as they were overwriting
the display.